+++ installer/cape2.sh
@@ -49,7 +49,7 @@ librenms_mdadm_enable=0
 librenms_megaraid_enable=0
 
 # disabling this will result in the web interface being disabled
-mongo_enable=1
+MONGO_ENABLE=1
 
 DIE_VERSION="3.07"
 
@@ -202,7 +202,7 @@ function librenms_sneck_config() {
 	else
 		echo "#clamav|/usr/lib/nagios/plugins/check_clamav -w $librenms_clamav_warn -c $librenms_clamav_crit"
 	fi
-	if [ "$mongo_enable" -ge 1 ]; then
+	if [ "$MONGO_ENABLE" -ge 1 ]; then
 		echo "mongodb|/usr/lib/nagios/plugins/check_mongodb.py $librenms_mongo_args"
 		echo 'cape_web_procs|/usr/lib/nagios/plugins/check_procs --ereg-argument-array "poetry.*bin/python manage.py" 1:'
 	else
@@ -602,12 +602,22 @@ function redsocks2() {
 }
 
 function distributed() {
-    sudo apt install uwsgi -y 2>/dev/null
-    sudo mkdir -p /data/{config,}db
-    sudo chown mongodb:mongodb /data/ -R
+    sudo apt install uwsgi uwsgi-plugin-python3 nginx -y 2>/dev/null
+    sudo -u ${USER} bash -c 'poetry run pip install flask flask-restful flask-sqlalchemy requests'
 
-    if [ ! -f /lib/systemd/system/mongos.service ]; then
-        cat >> /lib/systemd/system/mongos.service << EOL
+    sudo cp /opt/CAPEv2/uwsgi/capedist.ini /etc/uwsgi/apps-available/cape_dist.ini
+    sudo ln -s /etc/uwsgi/apps-available/cape_dist.ini /etc/uwsgi/apps-enabled
+
+    sudo -u postgres -H sh -c "psql -c \"CREATE DATABASE ${USER}dist\"";
+    sudo -u postgres -H sh -c "psql -d \"${USER}\" -c \"GRANT ALL PRIVILEGES ON DATABASE ${USER}dist to ${USER};\""
+    sudo -u postgres -H sh -c "psql -d \"${USER}\" -c \"ALTER DATABASE ${USER}dist OWNER TO ${USER};\""
+
+    if [ "$MONGO_ENABLE" -ge 1 ]; then
+        sudo mkdir -p /data/{config,}db
+        sudo chown mongodb:mongodb /data/ -R
+
+        if [ ! -f /lib/systemd/system/mongos.service ]; then
+            cat >> /lib/systemd/system/mongos.service << EOL
 [Unit]
 Description=Mongo shard service
 After=network.target
@@ -623,15 +633,16 @@ ExecStart=/usr/bin/mongos --configdb cape_config/${DIST_MASTER_IP}:27019 --port 
 [Install]
 WantedBy=multi-user.target
 EOL
-fi
-    systemctl daemon-reload
-    systemctl enable mongos.service
-    systemctl start mongos.service
+        fi
+        systemctl daemon-reload
+        systemctl enable mongos.service
+        systemctl start mongos.service
 
-    echo -e "\n\n\n[+] CAPE distributed documentation: https://github.com/kevoreilly/CAPEv2/blob/master/docs/book/src/usage/dist.rst"
-    echo -e "\t https://docs.mongodb.com/manual/tutorial/enable-authentication/"
-    echo -e "\t https://docs.mongodb.com/manual/administration/security-checklist/"
-    echo -e "\t https://docs.mongodb.com/manual/core/security-users/#sharding-security"
+        echo -e "\n\n\n[+] CAPE distributed documentation: https://github.com/kevoreilly/CAPEv2/blob/master/docs/book/src/usage/dist.rst"
+        echo -e "\t https://docs.mongodb.com/manual/tutorial/enable-authentication/"
+        echo -e "\t https://docs.mongodb.com/manual/administration/security-checklist/"
+        echo -e "\t https://docs.mongodb.com/manual/core/security-users/#sharding-security"
+    fi
 }
 
 function install_suricata() {
@@ -731,7 +742,7 @@ function install_yara() {
 }
 
 function install_mongo(){
-	if [ "$mongo_enable" -ge 1 ]; then
+	if [ "$MONGO_ENABLE" -ge 1 ]; then
 		echo "[+] Installing MongoDB"
 		# Mongo >=5 requires CPU AVX instruction support https://www.mongodb.com/docs/manual/administration/production-notes/#x86_64
 		if grep -q ' avx ' /proc/cpuinfo; then
@@ -885,13 +896,10 @@ function dependencies() {
 
     install_postgresql
 
-    # sudo su - postgres
-    #psql
     sudo -u postgres -H sh -c "psql -c \"CREATE USER ${USER} WITH PASSWORD '$PASSWD'\"";
     sudo -u postgres -H sh -c "psql -c \"CREATE DATABASE ${USER}\"";
     sudo -u postgres -H sh -c "psql -d \"${USER}\" -c \"GRANT ALL PRIVILEGES ON DATABASE ${USER} to ${USER};\""
     sudo -u postgres -H sh -c "psql -d \"${USER}\" -c \"ALTER DATABASE ${USER} OWNER TO ${USER};\""
-    #exit
 
     apt install apparmor-utils -y
     TCPDUMP_PATH=`which tcpdump`
@@ -1129,14 +1137,14 @@ function install_CAPE() {
     cp -r conf/*.conf "custom/conf"
     sed -i "/connection =/cconnection = postgresql://${USER}:${PASSWD}@localhost:5432/${USER}" custom/conf/cuckoo.conf
     # sed -i "/tor/{n;s/enabled = no/enabled = yes/g}" custom/conf/routing.conf
-    #sed -i "/memory_dump = off/cmemory_dump = on" custom/conf/cuckoo.conf
-    #sed -i "/machinery =/cmachinery = kvm" custom/conf/cuckoo.conf
+    # sed -i "/memory_dump = off/cmemory_dump = on" custom/conf/cuckoo.conf
+    # sed -i "/machinery =/cmachinery = kvm" custom/conf/cuckoo.conf
     sed -i "/interface =/cinterface = ${NETWORK_IFACE}" custom/conf/auxiliary.conf
 
     chown ${USER}:${USER} -R "/opt/CAPEv2/"
 
 	# default is enabled, so we only need to disable it
-	if [ "$mongo_enable" -lt 1 ]; then
+	if [ "$MONGO_ENABLE" -lt 1 ]; then
 		crudini --set custom/conf/reporting.conf mongodb enabled no
 	fi
 
@@ -1157,7 +1165,7 @@ function install_systemd() {
     cp /opt/CAPEv2/systemd/suricata.service /lib/systemd/system/suricata.service
     systemctl daemon-reload
 	cape_web_enable_string=''
-	if [ "$mongo_enable" -ge 1 ]; then
+	if [ "$MONGO_ENABLE" -ge 1 ]; then
 		cape_web_enable_string="cape-web"
 	fi
     systemctl enable cape cape-rooter cape-processor "$cape_web_enable_string" suricata