+++ changelog.md
@@ -1,3 +1,20 @@
+### [24.2.2023] CAPE 2.4: 🌻 Edition
+* New Unpacker option: `unpacker=2`
+* Deprecated:
+    * submitCAPE.py - no more additional jobs
+* Staging branch:
+    * We want to have CAPE stable. So new features will go to staging branch for 1-2 weeks before merged to master.
+    * If you want to help us to spot any possible issue use that branch on your dev side.
+* We need help to add as much tests as possible to cover all possible cases to prevent broken code.
+* Stop using `conf/` folder. All config should be in `custom/conf/`. This will simplify your life on CAPE updates when new entry added to base templates. [Details](https://github.com/kevoreilly/CAPEv2/blob/master/conf/readme.md)
+
+### [20.2.2023]
+* Scheduler update:
+    * A machine may be configured with `reserved = yes` in `<machinery>.conf`. For such machines, the scheduler will
+      not use it for tasks unless the user specifically requests it by its label.
+* Database update:
+    * The 'name' of all machines defined in `<machinery>.conf` must be unique. The same goes for their 'label' fields.
+
 ### [16.2.2023]
 * Monitor update: Hooking engine stability fix for detonation issues (e.g. Word)